Move policy parsing into separate module.
Separating the data input from the underlying software model/domain
creates new possibilities for loading data with different formats
from different sources.
