+++ tests/lib/aclgenerator_test.py
@@ -19,6 +19,7 @@ import unittest
 from lib import aclgenerator
 from lib import naming
 from lib import policy
+from lib import policyparser
 import mock
 
 
@@ -95,7 +96,7 @@ class ACLGeneratorTest(unittest.TestCase):
   def testEstablishedNostate(self):
     # When using "nostate" filter and a term with "option:: established"
     # have any protocol other than TCP and/or UDP should raise error.
-    pol = policy.ParsePolicy(GOOD_HEADER_1 + STATEFUL_ONLY_TERM, self.naming)
+    pol = policyparser.ParsePolicy(GOOD_HEADER_1 + STATEFUL_ONLY_TERM, self.naming)
     acl = ACLMock(pol, EXP_INFO)
     for _, terms in pol.filters:
       for term in terms:
@@ -104,7 +105,7 @@ class ACLGeneratorTest(unittest.TestCase):
 
   def testSupportedAF(self):
     # Unsupported address families should raise an error.
-    pol = policy.ParsePolicy(GOOD_HEADER_1 + GOOD_TERM_1, self.naming)
+    pol = policyparser.ParsePolicy(GOOD_HEADER_1 + GOOD_TERM_1, self.naming)
     acl = ACLMock(pol, EXP_INFO)
     for _, terms in pol.filters:
       for term in terms:
@@ -114,7 +115,7 @@ class ACLGeneratorTest(unittest.TestCase):
   def testTermNameBelowLimit(self):
     # Term name that is below specified limit should come out unchanged,
     # regardless of abbreviation and truncation settings.
-    pol = policy.ParsePolicy(GOOD_HEADER_1 + SHORT_TERM_NAME, self.naming)
+    pol = policyparser.ParsePolicy(GOOD_HEADER_1 + SHORT_TERM_NAME, self.naming)
     acl = ACLMock(pol, EXP_INFO)
     for _, terms in pol.filters:
       for term in terms:
@@ -130,7 +131,7 @@ class ACLGeneratorTest(unittest.TestCase):
   def testLongTermAbbreviation(self):
     # Term name that is above specified limit should come out abbreviated
     # when abbreviation is enabled.
-    pol = policy.ParsePolicy(GOOD_HEADER_1 + GOOD_LONG_TERM_NAME, self.naming)
+    pol = policyparser.ParsePolicy(GOOD_HEADER_1 + GOOD_LONG_TERM_NAME, self.naming)
     acl = ACLMock(pol, EXP_INFO)
     for _, terms in pol.filters:
       for term in terms:
@@ -141,7 +142,7 @@ class ACLGeneratorTest(unittest.TestCase):
   def testTermNameTruncation(self):
     # Term name that is above specified limit should come out truncated
     # when truncation is enabled.
-    pol = policy.ParsePolicy(GOOD_HEADER_1 + GOOD_LONG_TERM_NAME, self.naming)
+    pol = policyparser.ParsePolicy(GOOD_HEADER_1 + GOOD_LONG_TERM_NAME, self.naming)
     acl = ACLMock(pol, EXP_INFO)
     for _, terms in pol.filters:
       for term in terms:
@@ -151,7 +152,7 @@ class ACLGeneratorTest(unittest.TestCase):
   def testLongTermName(self):
     # Term name that is above specified limit and is impossible to abbreviate
     # should raise an exception.
-    pol = policy.ParsePolicy(GOOD_HEADER_1 + BAD_LONG_TERM_NAME, self.naming)
+    pol = policyparser.ParsePolicy(GOOD_HEADER_1 + BAD_LONG_TERM_NAME, self.naming)
     acl = ACLMock(pol, EXP_INFO)
     for _, terms in pol.filters:
       for term in terms:
